Extract IUserService.TwoFactorIsEnabledAsync into a new TwoFactorIsEnabledQuery class
